# src/lib/swagger/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "The LightSwitch API",
    "version": "1.0.0",
    "description": "API for managing projects, tasks, and users."
  },
  "servers": [
    {
      "url": "https://lightswitch.gitbook.dev/api"
    }
  ],
  "paths": {
...
    "/tasks": {
      "get": {
        "summary": "Get all tasks",
        "responses": {
          "200": {
            "description": "List of tasks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a new task",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Task"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Task created",
            "content": {
              "application/json": {
                "$ref": "#/components/schemas/Task"
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}": {
      "put": {
        "summary": "Update a task",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Task ID to update"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Task"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Task updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Task": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "projectId": { "type": "integer" },
          "title": { "type": "string" },
          "status": { "type": "string" }
        }
      },
...
    }
  }
}